*Compilation with RAD Studio 2010 *Added an OSThread::TerminateWait() function that initiates thread termination, then waits until the thread has terminated
git-svn-id: http://www.bci2000.org/svn/trunk@3470 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/shared/accessors/StateRef.h
@@ -191,7 +191,7 @@ const StateRefSigned&
 StateRefSigned::operator=( long inL )
 {
   const int bitsPerByte = 8;
-  if( mrStateRef->Length() == sizeof( 1L ) * bitsPerByte )
+  if( mrStateRef->Length() == static_cast<int>( sizeof( 1L ) * bitsPerByte ) )
     mrStateRef = inL;
   else
     mrStateRef = ( inL & ( ( 1L << mrStateRef->Length() ) - 1L ) );
